refactor(voting-api): move events to EventFactory
